+++ crates/oas3/src/error.rs
@@ -5,8 +5,6 @@ use std::io;
 use derive_more::derive::{Display, Error, From};
 
 use crate::spec::Error as SpecError;
-#[cfg(feature = "validation")]
-use crate::validation::Error as ValidationError;
 
 /// Top-level errors.
 #[derive(Debug, Display, Error, From)]
@@ -22,12 +20,4 @@ pub enum Error {
 
     #[display("Spec error")]
     Spec(SpecError),
-
-    #[cfg(feature = "validation")]
-    #[display("Validation error")]
-    Validation(ValidationError),
-
-    #[cfg(feature = "conformance")]
-    #[display("Reqwest error")]
-    Reqwest(reqwest::Error),
 }